requirements.txt
requests
python-whois
yara-python
